+++ projects/_03_independent_cqc/_02_clean/tests/jobs/test_clean_ind_cqc_filled_posts.py
@@ -45,6 +45,7 @@ class MainTests(CleanIndFilledPostsTests):
     @patch(f"{PATCH_PATH}.utils.write_to_parquet")
     @patch(f"{PATCH_PATH}.clean_capacity_tracker_non_res_outliers")
     @patch(f"{PATCH_PATH}.clean_capacity_tracker_care_home_outliers")
+    @patch(f"{PATCH_PATH}.forward_fill_latest_known_value")
     @patch(f"{PATCH_PATH}.clean_ascwds_filled_post_outliers")
     @patch(f"{PATCH_PATH}.cUtils.create_banded_bed_count_column")
     @patch(f"{PATCH_PATH}.cUtils.calculate_filled_posts_per_bed_ratio")
@@ -71,6 +72,7 @@ class MainTests(CleanIndFilledPostsTests):
         calculate_filled_posts_per_bed_ratio_mock: Mock,
         create_banded_bed_count_column_mock: Mock,
         clean_ascwds_filled_post_outliers_mock: Mock,
+        forward_fill_latest_known_value_mock: Mock,
         clean_capacity_tracker_care_home_outliers_mock: Mock,
         clean_capacity_tracker_non_res_outliers_mock: Mock,
         write_to_parquet_mock: Mock,
@@ -92,6 +94,7 @@ class MainTests(CleanIndFilledPostsTests):
         self.assertEqual(create_column_with_repeated_values_removed_mock.call_count, 2)
         self.assertEqual(calculate_filled_posts_per_bed_ratio_mock.call_count, 3)
         create_banded_bed_count_column_mock.assert_called_once()
+        forward_fill_latest_known_value_mock.assert_called_once()
         clean_ascwds_filled_post_outliers_mock.assert_called_once()
         clean_capacity_tracker_care_home_outliers_mock.assert_called_once()
         clean_capacity_tracker_non_res_outliers_mock.assert_called_once()
@@ -236,6 +239,12 @@ class MainTests(CleanIndFilledPostsTests):
 
         df = df.collect()
         self.assertEqual(df[0][IndCQC.number_of_beds], 1)
+
+    def test_days_to_repeat_forward_filling_is_correct(self):
+        self.assertEqual(
+            job.NumericalValues.number_of_days_to_forward_fill,
+            65,
+        )
 
 
 class CalculateTimeRegisteredForTests(CleanIndFilledPostsTests):